Better name for the matrix testing functions.
