+++ src/gates/single/pauli.rs
@@ -203,7 +203,7 @@ pub fn z(qubit: &mut Qubit) -> &mut Qubit {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::gates::test_helpers::assert_matrix_result;
+    use crate::gates::matrix_testing::assert_matrix_result;
     use num::complex::Complex64;
     use num::complex::ComplexFloat;
     use rstest::rstest;
@@ -240,7 +240,7 @@ mod tests {
     ) {
         // Verify against matrix operations first to validate expected output
         assert_matrix_result(input_coeffs, expected_coeffs, || {
-            crate::gates::matrix_operations::x()
+            crate::gates::matrix_testing::x()
         });
 
         let (alpha_in, beta_in) = input_coeffs;
@@ -300,7 +300,7 @@ mod tests {
     ) {
         // Verify against matrix operations first to validate expected output
         assert_matrix_result(input_coeffs, expected_coeffs, || {
-            crate::gates::matrix_operations::y()
+            crate::gates::matrix_testing::y()
         });
 
         let (alpha_in, beta_in) = input_coeffs;
@@ -360,7 +360,7 @@ mod tests {
     ) {
         // Verify against matrix operations first to validate expected output
         assert_matrix_result(input_coeffs, expected_coeffs, || {
-            crate::gates::matrix_operations::z()
+            crate::gates::matrix_testing::z()
         });
 
         let (alpha_in, beta_in) = input_coeffs;
